requirements.txt
streamlit
openai
youtube-transcript-api
yt-dlp
whisper
python-dotenv

